% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-functions.R
\name{area_hexagon}
\alias{area_hexagon}
\title{Area of a Hexagon}
\usage{
area_hexagon(x)
}
\arguments{
\item{x}{Length of one side of the hexagon. (positive number)}
}
\value{
The area of a hexagon with side lengths x.
}
\description{
A simple function to find the area of a hexagon
}
\examples{
area_hexagon(3)
}
