% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_nonagon.R
\name{area_nonagon}
\alias{area_nonagon}
\title{Area of a Nonagon}
\usage{
area_nonagon(length, graph)
}
\arguments{
\item{length}{Length of the regular nonagon (positive number)}

\item{graph}{0 or 1, with 1 requesting a graph and 0 not (indicator variable)}
}
\value{
The area of a nonagon with given length and a graph of the square if indicated
}
\description{
A simple function to find the area of a regular nonagon
}
\examples{
area_nonagon(5, 0)
area_nonagon(5, 1)
}
