% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_octagon.R
\name{area_octagon}
\alias{area_octagon}
\title{Area of a Regular Octagon}
\usage{
area_octagon(x, graph)
}
\arguments{
\item{x}{Length of one side of the regular octagon (positive number)}

\item{graph}{Indicator variable for graphing option (0 for no graph, 1 for graph)}
}
\value{
The area of a regular octagon with side lengths x.
}
\description{
A simple function to find the area of a regular octagon
}
\examples{
area_octagon(3, 0 )
}
