% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-functions.R
\name{area_octogon}
\alias{area_octogon}
\title{Area of a Octogon}
\usage{
area_octogon(x)
}
\arguments{
\item{x}{Length of one side of the octogon (positive number)}
}
\value{
The area of a octogon with side lengths x.
}
\description{
A simple function to find the area of a octogon
}
\examples{
area_octogon(3)
}
