% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_square.R
\name{area_rectangle}
\alias{area_rectangle}
\title{Area of a Rectangle}
\usage{
area_rectangle(x, y, graph)
}
\arguments{
\item{x}{Width of the rectangle (positive number)}

\item{y}{Height of the rectangle (positive number)}

\item{graph}{0 or 1, with 1 requesting a graph and 0 not (indicator variable)}
}
\value{
The area of a rectangle with width x and height y, and a graph of the square if requested.
}
\description{
A simple function to find the area of a rectangle.
}
\examples{
area_rectangle(5,4, 0)
area_rectangle(5,4, 1)
}
