% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_triangle.R
\name{area_triangle}
\alias{area_triangle}
\title{Area of a Triangle}
\usage{
area_triangle(base, height, graph)
}
\arguments{
\item{base}{Length of the base of the triangle (positive number)}

\item{height}{Height of the triangle, which creates a 90 degree angle with the base (positive number)}

\item{graph}{Indicator variable to graph (0 correlates to not creating a graph, 1 is to graph an isosceles triangle, and 2 is to graph a right triangle)}
}
\value{
The area of a triangle with a base of x and height of y, along with a graph if indicated
}
\description{
A simple function to find the area of a triangle
}
\examples{
area_triangle(2, 4, 0)
area_triangle(2, 4, 1)
area_triangle(2, 4, 2)
}
