% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule.R
\name{capsule}
\alias{capsule}
\title{Calculates the surface area, or volume, or both of a capsule with the provided dimension}
\usage{
capsule(side_len, radius, return = "both")
}
\arguments{
\item{side_len}{numeric flat side length dimension}

\item{radius}{numeric radius of hemisphere shaped part of capsule dimension}

\item{return}{string input desired return value(s) where the default is "both" volume and surface area}
}
\value{
surface area, or volume, or both of capsule
}
\description{
Calculates the surface area, or volume, or both of a capsule with the provided dimension
}
