% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_volume.R
\name{capsule_volume}
\alias{capsule_volume}
\title{Calculates the volume of a capsule with the provided dimensions}
\usage{
capsule_volume(side_len, radius)
}
\arguments{
\item{side_len}{side length dimension}

\item{radius}{radius dimension}
}
\value{
volume of capsule
}
\description{
Calculates the volume of a capsule with the provided dimensions
}
