% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{cube}
\alias{cube}
\title{Calculates the volume of a cube with the provide dimension}
\usage{
cube(length)
}
\arguments{
\item{length}{length dimension}
}
\value{
volume of cube
}
\description{
Calculates the volume of a cube with the provide dimension
}
