% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{cylinder}
\alias{cylinder}
\title{Calculates the volume of a cylinder with the provide dimensions}
\usage{
cylinder(radius, height)
}
\arguments{
\item{radius}{radius dimension}

\item{height}{height dimension}
}
\value{
volume of cylinder
}
\description{
Calculates the volume of a cylinder with the provide dimensions
}
