% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{cylinder_volume}
\alias{cylinder_volume}
\title{Calculates the volume of a cylinder with the provided dimensions}
\usage{
cylinder_volume(base_radius, height)
}
\arguments{
\item{base_radius}{radius of the circular base}

\item{height}{height dimension}
}
\value{
volume of cylinder
}
\description{
Calculates the volume of a cylinder with the provided dimensions
}
