% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{ellipsoid_volume}
\alias{ellipsoid_volume}
\title{Calculates the volume of an ellipsoid with the provided dimensions}
\usage{
ellipsoid_volume(semi_axe_1, semi_axe_2, semi_axe_3)
}
\arguments{
\item{semi_axe_1}{1st semi-axe dimension}

\item{semi_axe_2}{2nd semi-axe dimension}

\item{semi_axe_3}{3rd semi-axe dimension}
}
\value{
volume of ellipsoid
}
\description{
Calculates the volume of an ellipsoid with the provided dimensions
}
