% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_hexagon.R
\name{graph_hexagon}
\alias{graph_hexagon}
\title{Graph Hexagon}
\usage{
graph_hexagon(x)
}
\arguments{
\item{x}{Length of one side of the regular hexagon. (positive number)}
}
\value{
Graph of the area of a hexagon
}
\description{
A simple function to graph a hexagon.
}
\examples{
graph_hexagon(2)
}
