% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_octagon.R
\name{graph_octagon}
\alias{graph_octagon}
\title{Graph Octagon}
\usage{
graph_octagon(x)
}
\arguments{
\item{x}{Length of the octagon (positive number)}
}
\value{
Graph of the area of an octagon
}
\description{
A simple function to graph an octagon.
}
\examples{
graph_octagon(2)
}
