% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pentagon.R
\name{graph_pentagon}
\alias{graph_pentagon}
\title{Graph Pentagon}
\usage{
graph_pentagon(length, apothem)
}
\arguments{
\item{length}{Length of one side of the regular pentagon, also known as a perimeter. (positive number)}

\item{apothem}{Length from the center of one side of the regular pentagon to the center, also known as the apothem. (positive number)}
}
\value{
Graph of the area of a pentagon
}
\description{
A simple function to graph a pentagon
}
\examples{
graph_pentagon(2, 4)
}
