% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_square.R
\name{graph_square}
\alias{graph_square}
\title{Graph Square}
\usage{
graph_square(x, y)
}
\arguments{
\item{x}{Width of the square (positive number)}

\item{y}{Height of the square (positive number)}
}
\value{
Graph of the area of a square
}
\description{
A simple function to find the area of a square.
}
\examples{
graph_square(2, 4)
}
