% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_trapezoid.R
\name{graph_trapezoid}
\alias{graph_trapezoid}
\title{Graph Trapezoid}
\usage{
graph_trapezoid(x, y, z)
}
\arguments{
\item{x}{Length of the smaller base of the regular trapezoid. (positive number)}

\item{y}{Length of the larger base of the regular trapezoid. (positive number)}

\item{z}{Height of the regular trapezoid, which creates a 90 degree angle between both bases. (positive number)}
}
\value{
Graph of the area of a trapezoid
}
\description{
A simple function to graph a trapezoid.
}
\examples{
graph_trapezoid(2, 4, 3)
}
