% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_triangle.R
\name{graph_triangle}
\alias{graph_triangle}
\title{Graph Triangle}
\usage{
graph_triangle(x, y, type)
}
\arguments{
\item{x}{Width of the traingle (positive number)}

\item{y}{Height of the triangle (positive number)}

\item{type}{Either an isoscoles or right triangle to graph (1,2)}
}
\value{
Graph of the area of a triangle
}
\description{
A simple function to graph a triangle.
}
\examples{
graph_triangle(2, 4, 1)
}
