% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemisphere.R
\name{hemisphere}
\alias{hemisphere}
\title{Calculates the surface area, or volume, or both of a semi-sphere with the provided dimension}
\usage{
hemisphere(base_radius, return = "both")
}
\arguments{
\item{base_radius}{base radius dimension}

\item{return}{string input desired return value(s) where the default is "both" volume and surface area}
}
\value{
surface area, or volume, or both of hemisphere
}
\description{
Calculates the surface area, or volume, or both of a semi-sphere with the provided dimension
}
