% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{hemisphere_volume}
\alias{hemisphere_volume}
\title{Calculates the volume of a semi-sphere with the provided dimension}
\usage{
hemisphere_volume(base_radius)
}
\arguments{
\item{base_radius}{base radius dimension}
}
\value{
volume of hemisphere
}
\description{
Calculates the volume of a semi-sphere with the provided dimension
}
