% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\docType{data}
\name{primes}
\alias{primes}
\title{Prime Numbers}
\format{
\subsection{\code{primes}}{

A data frame with 1,000 rows and 2 columns:
\describe{
\item{n}{The chronological prime number order}
\item{prime_n}{Prime number}
}
}
}
\source{
\url{https://www.di-mgt.com.au/primes1000.html}
}
\usage{
primes
}
\description{
The first 1000 prime numbers.
}
\keyword{datasets}
