% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{prism}
\alias{prism}
\title{Calculates the volume of a prism with the provide dimensions}
\usage{
prism(base, height)
}
\arguments{
\item{base}{base dimension}

\item{height}{height dimension}
}
\value{
volume of prism
}
\description{
Calculates the volume of a prism with the provide dimensions
}
