% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_volume.R
\name{prism_volume}
\alias{prism_volume}
\title{Calculates the volume of a prism with the provided dimensions}
\usage{
prism_volume(base_area, height)
}
\arguments{
\item{base_area}{area of base}

\item{height}{height dimension}
}
\value{
volume of prism
}
\description{
Calculates the volume of a prism with the provided dimensions
}
