% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular_prism.R
\name{rectangular_prism}
\alias{rectangular_prism}
\title{Calculates the surface area, or volume, or bothof a rectangular solid with the provided dimensions}
\usage{
rectangular_prism(length, width, height, return = "both")
}
\arguments{
\item{length}{length dimension}

\item{width}{width dimension}

\item{height}{height dimension}

\item{return}{string input desired return value(s) where the default is "both" volume and surface area}
}
\description{
Calculates the surface area, or volume, or bothof a rectangular solid with the provided dimensions
}
