% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{rectangular_solid_volume}
\alias{rectangular_solid_volume}
\title{Calculates the volume of a rectangular solid with the provided dimensions}
\usage{
rectangular_solid_volume(length, width, height)
}
\arguments{
\item{length}{length dimension}

\item{width}{width dimension}

\item{height}{height dimension}
}
\value{
volume of rectangular solid
}
\description{
Calculates the volume of a rectangular solid with the provided dimensions
}
