% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_circular_cone_volume.R
\name{right_circular_cone_volume}
\alias{right_circular_cone_volume}
\title{Calculates the volume of a right circular cone with the provided dimensions}
\usage{
right_circular_cone_volume(base_radius, height)
}
\arguments{
\item{base_radius}{radius of the circular base}

\item{height}{height dimension}
}
\value{
volume of right circular cone
}
\description{
Calculates the volume of a right circular cone with the provided dimensions
}
