% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{sphere}
\alias{sphere}
\title{Calculates the volume of a sphere with the provide dimensions}
\usage{
sphere(radius)
}
\arguments{
\item{radius}{radius dimension}
}
\value{
volume of sphere
}
\description{
Calculates the volume of a sphere with the provide dimensions
}
