% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_pyramid.R
\name{square_pyramid}
\alias{square_pyramid}
\title{Calculates the surface area, or volume, or both of a square pyramid with the provided dimension}
\usage{
square_pyramid(side_len, height, return = "both")
}
\arguments{
\item{side_len}{side length of base}

\item{height}{height dimension}

\item{return}{string input desired return value(s) where the default is "both" volume and surface area}
}
\value{
surface area, or volume, or both of square pyramid
}
\description{
Calculates the surface area, or volume, or both of a square pyramid with the provided dimension
}
