% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_pyramid_volume.R, R/volume-functions.R
\name{square_pyramid_volume}
\alias{square_pyramid_volume}
\title{Calculates the volume of a square pyramid with the provided dimensions}
\usage{
square_pyramid_volume(base_area, height)

square_pyramid_volume(base_area, height)
}
\arguments{
\item{base_area}{area of base}

\item{height}{height dimension}
}
\value{
volume of square pyramid

volume of square pyramid
}
\description{
Calculates the volume of a square pyramid with the provided dimensions

Calculates the volume of a square pyramid with the provided dimensions
}
