% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_capsule.R
\name{surf_area_capsule}
\alias{surf_area_capsule}
\title{Calculates the surface area of capsule with the provided dimension}
\usage{
surf_area_capsule(side_len, radius)
}
\arguments{
\item{side_len}{flat side length dimension}

\item{radius}{radius of hemisphere shaped part of capsule dimension}
}
\value{
surface area of capsule
}
\description{
Calculates the surface area of capsule with the provided dimension
}
