% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_cube.R
\name{surf_area_cube}
\alias{surf_area_cube}
\title{Calculates the surface area of a cube with the provided dimensions}
\usage{
surf_area_cube(side_len)
}
\arguments{
\item{side_len}{side length dimension}
}
\value{
surface area of a cube
}
\description{
Calculates the surface area of a cube with the provided dimensions
}
