% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_cylinder.R
\name{surf_area_cylinder}
\alias{surf_area_cylinder}
\title{Calculates the surface area of cylinder with the provided dimensions}
\usage{
surf_area_cylinder(base_radius, height)
}
\arguments{
\item{base_radius}{radius of the circular base dimension}

\item{height}{height dimension}
}
\value{
surface area of cylinder
}
\description{
Calculates the surface area of cylinder with the provided dimensions
}
