% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_hemisphere.R
\name{surf_area_hemisphere}
\alias{surf_area_hemisphere}
\title{Calculates the surface area of a semi-sphere with the provided dimension}
\usage{
surf_area_hemisphere(base_radius)
}
\arguments{
\item{base_radius}{base radius dimension}
}
\value{
surface area of hemisphere
}
\description{
Calculates the surface area of a semi-sphere with the provided dimension
}
