% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_rectangular_volume.R
\name{surf_area_rectangular_solid}
\alias{surf_area_rectangular_solid}
\title{Calculates the surface area of a rectangular solid with the provided dimensions}
\usage{
surf_area_rectangular_solid(length, width, height)
}
\arguments{
\item{length}{length dimension}

\item{width}{width dimension}

\item{height}{height dimension}
}
\value{
surface area of rectangular solid
}
\description{
Calculates the surface area of a rectangular solid with the provided dimensions
}
