% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_sphere.R
\name{surf_area_sphere}
\alias{surf_area_sphere}
\title{Calculates the surface area of a sphere with the provided dimension}
\usage{
surf_area_sphere(radius)
}
\arguments{
\item{radius}{radius dimension}
}
\value{
surface area of sphere
}
\description{
Calculates the surface area of a sphere with the provided dimension
}
