% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_square_pyramid.R
\name{surf_area_square_pyramid}
\alias{surf_area_square_pyramid}
\title{Calculates the surface area of a square pyramid with the provided dimensions}
\usage{
surf_area_square_pyramid(side_len, height)
}
\arguments{
\item{side_len}{side length of base}

\item{height}{height dimension}
}
\value{
volume of square pyramid
}
\description{
Calculates the surface area of a square pyramid with the provided dimensions
}
