% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area_tetrahedron.R
\name{surf_area_tetrahedron}
\alias{surf_area_tetrahedron}
\title{Calculates the surface area of a tetrahedron with the provided dimensions}
\usage{
surf_area_tetrahedron(length)
}
\arguments{
\item{length}{length of the edge}
}
\value{
surface area of tetrahedron
}
\description{
Calculates the surface area of a tetrahedron with the provided dimensions
}
