% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube_volume.R
\name{tube_volume}
\alias{tube_volume}
\title{Calculates the volume of a tube with the provided dimensions}
\usage{
tube_volume(outer_radius, inner_radius, height)
}
\arguments{
\item{outer_radius}{outer radius dimension}

\item{inner_radius}{inner radius dimension}

\item{height}{height dimension}
}
\value{
volume of tube
}
\description{
Calculates the volume of a tube with the provided dimensions
}
