% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_capsule.R
\name{volume_capsule}
\alias{volume_capsule}
\title{Calculates the volume of a capsule with the provided dimensions}
\usage{
volume_capsule(side_len, radius)
}
\arguments{
\item{side_len}{side length dimension}

\item{radius}{radius dimension}
}
\value{
volume of capsule
}
\description{
Calculates the volume of a capsule with the provided dimensions
}
