% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_cube.R
\name{volume_cube}
\alias{volume_cube}
\title{Calculates the volume of a cube with the provided dimension}
\usage{
volume_cube(length)
}
\arguments{
\item{length}{length of edge or side}
}
\value{
volume of cube
}
\description{
Calculates the volume of a cube with the provided dimension
}
