% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_cylinder.R
\name{volume_cylinder}
\alias{volume_cylinder}
\title{Calculates the volume of a cylinder with the provided dimensions}
\usage{
volume_cylinder(base_radius, height)
}
\arguments{
\item{base_radius}{radius of the circular base}

\item{height}{height dimension}
}
\value{
volume of cylinder
}
\description{
Calculates the volume of a cylinder with the provided dimensions
}
