% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_hemisphere.R
\name{volume_hemisphere}
\alias{volume_hemisphere}
\title{Calculates the volume of a semi-sphere with the provided dimension}
\usage{
volume_hemisphere(base_radius)
}
\arguments{
\item{base_radius}{base radius dimension}
}
\value{
volume of hemisphere
}
\description{
Calculates the volume of a semi-sphere with the provided dimension
}
