% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_prism.R
\name{volume_prism}
\alias{volume_prism}
\title{Calculates the volume of a prism with the provided dimensions}
\usage{
volume_prism(base_area, height)
}
\arguments{
\item{base_area}{area of base}

\item{height}{height dimension}
}
\value{
volume of prism
}
\description{
Calculates the volume of a prism with the provided dimensions
}
