% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_rectangular_solid.R
\name{volume_rectangular_solid}
\alias{volume_rectangular_solid}
\title{Calculates the volume of a rectangular solid with the provided dimensions}
\usage{
volume_rectangular_solid(length, width, height)
}
\arguments{
\item{length}{length dimension}

\item{width}{width dimension}

\item{height}{height dimension}
}
\value{
volume of rectangular solid
}
\description{
Calculates the volume of a rectangular solid with the provided dimensions
}
