% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_right_circular_cone.R
\name{volume_right_circular_cone}
\alias{volume_right_circular_cone}
\title{Calculates the volume of a right circular cone with the provided dimensions}
\usage{
volume_right_circular_cone(base_radius, height)
}
\arguments{
\item{base_radius}{radius of the circular base}

\item{height}{height dimension}
}
\value{
volume of right circular cone
}
\description{
Calculates the volume of a right circular cone with the provided dimensions
}
