% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_sphere.R
\name{volume_sphere}
\alias{volume_sphere}
\title{Calculates the volume of a sphere with the provided dimension}
\usage{
volume_sphere(radius)
}
\arguments{
\item{radius}{radius dimension}
}
\value{
volume of sphere
}
\description{
Calculates the volume of a sphere with the provided dimension
}
