% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_square_pyramid.R
\name{volume_square_pyramid}
\alias{volume_square_pyramid}
\title{Calculates the volume of a square pyramid with the provided dimensions}
\usage{
volume_square_pyramid(base_area, height)
}
\arguments{
\item{base_area}{area of base}

\item{height}{height dimension}
}
\value{
volume of square pyramid
}
\description{
Calculates the volume of a square pyramid with the provided dimensions
}
