% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_tetrahedron.R
\name{volume_tetrahedron}
\alias{volume_tetrahedron}
\title{Calculates the volume of a tetrahedron with the provided dimensions}
\usage{
volume_tetrahedron(length)
}
\arguments{
\item{length}{length of the edge}
}
\value{
volume of tetrahedron
}
\description{
Calculates the volume of a tetrahedron with the provided dimensions
}
