% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Area Functions.R
\name{area_circle}
\alias{area_circle}
\title{Area of a Circle}
\usage{
area_circle(x)
}
\arguments{
\item{x}{Radius of the circle (positive number)}
}
\value{
The area of the circle with radius x.
}
\description{
A simple function to find the area of a circle
}
\examples{
area_circle(7)
}
