% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_nonagon.R
\name{area_nonagon}
\alias{area_nonagon}
\title{Area of a Nonagon}
\usage{
area_nonagon(length)
}
\arguments{
\item{length}{Length of one side of the regular nonagon (positive number)}
}
\value{
The area of a nonagon with given length and a graph of the square if indicated
}
\description{
A simple function to find the area of a regular nonagon
}
\examples{
area_nonagon(5)
}
