% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-functions.R
\name{area_octagon}
\alias{area_octagon}
\title{Area of a Regular Octagon}
\usage{
area_octagon(x)
}
\arguments{
\item{x}{Length of one side of the regular octagon (positive number)}
}
\value{
The area of a regular octagon with side lengths x.
}
\description{
A simple function to find the area of a regular octagon
}
\examples{
area_octagon(3)
}
