% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_pentagon.R
\name{area_pentagon}
\alias{area_pentagon}
\title{Area of a Regular Pentagon}
\usage{
area_pentagon(length, apothem, graph)
}
\arguments{
\item{length}{Length of one side of the regular pentagon, also known as a perimeter (positive number)}

\item{apothem}{Length from the center of one side of the regular pentagon to the center, also known as the apothem (positive number)}

\item{graph}{Indicator variable for graphing option (0 for no graph, 1 for graph)}
}
\value{
The area of a regular pentagon with a perimeter x and apothem y, with option of graphing
}
\description{
A simple function to find the area of a regular pentagon
}
\examples{
area_pentagon(2, 3, 0)
}
