% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-functions.R
\name{area_square}
\alias{area_square}
\title{Area of a Square}
\usage{
area_square(x, y)
}
\arguments{
\item{x}{Width of the square (positive number)}

\item{y}{Height of the square (positive number)}
}
\value{
The area of a square with width x and height y.
}
\description{
A simple function to find the area of a square.
}
\examples{
area_square(5,4)
}
