% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_trapezoid.R
\name{area_trapezoid}
\alias{area_trapezoid}
\title{Area of a Regular Trapezoid}
\usage{
area_trapezoid(x, y, z)
}
\arguments{
\item{x}{Length of the smaller base of the regular trapezoid. (positive number)}

\item{y}{Length of the larger base of the regular trapezoid. (positive number)}

\item{z}{Height of the regular trapezoid, which creates a 90 degree angle between both bases. (positive number)}
}
\value{
The area of a regular trapezoid with a bases of x and y and height of z.
}
\description{
A simple function to find the area of a regular trapezoid
}
\examples{
area_trapezoid(2, 3, 4)
}
