% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-functions.R
\name{area_triangle}
\alias{area_triangle}
\title{Area of a Triangle}
\usage{
area_triangle(x, y)
}
\arguments{
\item{x}{Length of the base of the triangle. (posotive number)}

\item{y}{Height of the triangle, which creates a 90 degree angle with the base. (positive number)}
}
\value{
The area of a triangle with a base of x and height of y.
}
\description{
A simple function to find the area of a triangle.
}
\examples{
area_triangle(2, 4)
}
