% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Calculates the surface area, or volume, or both of a cube with the provided dimension}
\usage{
cube(length, return = "both")
}
\arguments{
\item{length}{length of edge or side}

\item{return}{string input desired return value(s) where the default is "both" volume and surface area}
}
\value{
surface area, or volume, or both of cube
}
\description{
Calculates the surface area, or volume, or both of a cube with the provided dimension
}
