% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-functions.R, R/cube_volume.R
\name{cube_volume}
\alias{cube_volume}
\title{Calculates the volume of a cube with the provided dimension}
\usage{
cube_volume(length)

cube_volume(length)
}
\arguments{
\item{length}{length of edge or side}
}
\value{
volume of cube

volume of cube
}
\description{
Calculates the volume of a cube with the provided dimension

Calculates the volume of a cube with the provided dimension
}
