% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_circle.R
\name{graph_circle}
\alias{graph_circle}
\title{Graph Circle}
\usage{
graph_circle(x)
}
\arguments{
\item{x}{Radius of the circle (positive number)}
}
\value{
The graph of a circle with radius x.
}
\description{
A simple function to graph a circle.
}
\examples{
graph_circle(2)
}
