% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\docType{data}
\name{primes}
\alias{primes}
\title{Primes Dataset}
\format{
\subsection{\code{primes}}{

A data frame with 1,000 rows and 2 columns:
\describe{
\item{index}{Index of the prime number}
\item{prime_n}{Prime number}
\item{type}{If the prime number is even or odd}
}
}
}
\source{
\url{https://www.di-mgt.com.au/primes1000.html}
}
\usage{
primes
}
\description{
A dataset with the first 1000 prime numbers.
}
\keyword{datasets}
