% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere}
\alias{sphere}
\title{Calculates the surface area, or volume, or both of a sphere with the provided dimension}
\usage{
sphere(radius, return = "both")
}
\arguments{
\item{radius}{radius dimension}

\item{return}{string input desired return value(s) where the default is "both" volume and surface area}
}
\value{
surface area, or volume, or both of sphere
}
\description{
Calculates the surface area, or volume, or both of a sphere with the provided dimension
}
