% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-functions.R
\name{sphere_volume}
\alias{sphere_volume}
\title{Calculates the volume of a sphere with the provided dimension}
\usage{
sphere_volume(radius)
}
\arguments{
\item{radius}{radius dimension}
}
\value{
volume of sphere
}
\description{
Calculates the volume of a sphere with the provided dimension
}
