% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrahedron_volume.R
\name{tetrahedron_volume}
\alias{tetrahedron_volume}
\title{Calculates the volume of a tetrahedron with the provided dimensions}
\usage{
tetrahedron_volume(length)
}
\arguments{
\item{length}{length of the edge}
}
\value{
volume of tetrahedron
}
\description{
Calculates the volume of a tetrahedron with the provided dimensions
}
